fix: resolve compilation errors and update v2.3 documentation
- Fix PagedKVCache type by adding type alias to PagedAttention
- Add Debug derive to PageTable and PagedAttention structs
- Fix sha2 dependency placement in Cargo.toml
- Fix duplicate ModelInfo/TaskType exports with aliases
- Fix type cast in upload.rs parameters method

Documentation:
- Update RuvLLM crate README to v2.3 with new features
- Add npm package README with API reference
- Update issue #118 with RuvLTRA-Medium, LoRA adapters, Hub integration

v2.3 Features documented:
- RuvLTRA-Medium 3B model
- HuggingFace Hub integration
- 5 task-specific LoRA adapters
- Adapter merging (TIES, DARE, SLERP)
- Hot-swap adapter management
- Claude dataset training system

+++ crates/ruvllm/src/hub/upload.rs
@@ -278,7 +278,7 @@ impl ModelUploader {
             .task(TaskType::TextGeneration)
             .framework(Framework::Gguf)
             .architecture(&metadata.architecture)
-            .parameters(metadata.params_b * 1e9)
+            .parameters((metadata.params_b * 1e9) as u64)
             .context_length(metadata.context_length);
 
         if let Some(quant) = &metadata.quantization {
